+++ test/hlx_server_test.js
@@ -99,7 +99,7 @@ describe('Helix Server', () => {
     }
   });
 
-  it('proxy blobs requests to azure blob storage', async () => {
+  it('proxy blobs requests via inner cdn', async () => {
     const cwd = await setupProject(path.join(SPEC_ROOT, 'local'), testRoot);
     const project = new HelixProject()
       .withCwd(cwd)
@@ -108,12 +108,9 @@ describe('Helix Server', () => {
       .withLogsDir(path.resolve(cwd, 'logs'));
     await project.init();
 
-    nock('https://hlx.blob.core.windows.net')
-      .get('/external/098af326aa856bb42ce9a21240cf73d6f64b0b45')
-      .reply(function handler() {
-        assert.equal(this.req.headers.host, 'hlx.blob.core.windows.net');
-        return [304, '', {}];
-      });
+    nock('https://master--default--local.hlx.page')
+      .get('/hlx_098af326aa856bb42ce9a21240cf73d6f64b0b45.png')
+      .reply(304);
 
     try {
       await project.start();
@@ -123,7 +120,7 @@ describe('Helix Server', () => {
     }
   });
 
-  it('proxy fonts requests to typekit CDN', async () => {
+  it('proxy fonts requests via inner cdn', async () => {
     const cwd = await setupProject(path.join(SPEC_ROOT, 'local'), testRoot);
     const project = new HelixProject()
       .withCwd(cwd)
@@ -132,77 +129,39 @@ describe('Helix Server', () => {
       .withLogsDir(path.resolve(cwd, 'logs'));
     await project.init();
 
-    nock('https://use.typekit.net')
-      .get('/pnv6nym.css')
-      .reply(() => [200, '', {}]);
+    nock('https://master--default--local.hlx.page')
+      .get('/hlx_fonts/af/507e89/00000000000000003b9aee49/27/l?primer=7cdcb44be4a7db8877ffa5c0007b8dd865b3bbc383831fe2ea177f62257a9191&fvd=i4&v=3')
+      .reply(200);
 
     try {
       await project.start();
-      await assertHttp(`http://localhost:${project.server.port}/hlx_fonts/pnv6nym.css`, 200);
+      await assertHttp(`http://localhost:${project.server.port}/hlx_fonts/af/507e89/00000000000000003b9aee49/27/l?primer=7cdcb44be4a7db8877ffa5c0007b8dd865b3bbc383831fe2ea177f62257a9191&fvd=i4&v=3`, 200);
     } finally {
       await project.stop();
     }
   });
 
-  // currently not possible to test with helix-fetch
-  it.skip('proxy query requests to algolia endpoint', async () => {
+  it('proxy query requests to inner cdn', async () => {
     const cwd = await setupProject(path.join(SPEC_ROOT, 'local'), testRoot);
 
     const idxCfg = new IndexConfig()
       .withConfigPath(path.resolve(path.join(SPEC_ROOT, 'local'), 'helix-index.yaml'));
     await idxCfg.init();
 
-    const algoliaAppID = 'fake-id';
-    const algoliaAPIKey = 'fake-key';
-
     const project = new HelixProject()
       .withCwd(cwd)
       .withBuildDir('./build')
       .withHttpPort(0)
-      .withIndexConfig(idxCfg)
-      .withAlgoliaAppID(algoliaAppID)
-      .withAlgoliaAPIKey(algoliaAPIKey)
       .withLogsDir(path.resolve(cwd, 'logs'));
     await project.init();
 
-    function handler() {
-      if (this.req.headers['x-algolia-api-key'] === algoliaAPIKey
-        && this.req.headers['x-algolia-application-id'] === algoliaAppID) {
-        return [200, '', {}];
-      }
-      return [403, '', {}];
-    }
-
-    nock(`https://${algoliaAppID}-dsn.algolia.net`)
-      .get('/1/indexes/local--default--blog-posts?query=*&hitsPerPage=25')
-      .reply(handler);
+    nock('https://master--default--local.hlx.page')
+      .get('/_query/blog-posts/all')
+      .reply(200);
 
     try {
       await project.start();
       await assertHttp(`http://localhost:${project.server.port}/_query/blog-posts/all`, 200);
-    } finally {
-      await project.stop();
-    }
-  });
-
-  it('proxies query requests to algolia (no index configuration found)', async () => {
-    const cwd = await setupProject(path.join(SPEC_ROOT, 'local'), testRoot);
-
-    const algoliaAppID = 'fake-id';
-    const algoliaAPIKey = 'fake-key';
-
-    const project = new HelixProject()
-      .withCwd(cwd)
-      .withBuildDir('./build')
-      .withHttpPort(0)
-      .withAlgoliaAppID(algoliaAppID)
-      .withAlgoliaAPIKey(algoliaAPIKey)
-      .withLogsDir(path.resolve(cwd, 'logs'));
-    await project.init();
-
-    try {
-      await project.start();
-      await assertHttp(`http://localhost:${project.server.port}/_query/blog-posts/all`, 404);
     } finally {
       await project.stop();
     }
@@ -963,7 +922,7 @@ describe('Content Proxy Tests', () => {
     await project.init();
 
     nock('https://adobeioruntime.net')
-      .get('/api/v1/web/helix/helix-services/content-proxy@v1?owner=local&repo=default&path=%2Fms%2Ffoo.docx&ignore=github')
+      .get('/api/v1/web/helix/helix-services/content-proxy@v2?owner=local&repo=default&path=%2Fms%2Ffoo.docx&ignore=github')
       .reply(200, '# Hello\n\ndocx\n');
     try {
       await project.start();
@@ -996,7 +955,7 @@ describe('Content Proxy Tests', () => {
     await project.init();
 
     nock('https://adobeioruntime.net')
-      .get('/api/v1/web/helix/helix-services/content-proxy@v1')
+      .get('/api/v1/web/helix/helix-services/content-proxy@v2')
       .query({
         owner: 'local',
         repo: 'default',
@@ -1026,7 +985,7 @@ describe('Content Proxy Tests', () => {
     await project.init();
 
     nock('https://adobeioruntime.net')
-      .get('/api/v1/web/helix/helix-services/content-proxy@v1')
+      .get('/api/v1/web/helix/helix-services/content-proxy@v2')
       .query({
         owner: 'local',
         repo: 'default',